Renamed  `ICSQuery ` to  `CalendarQuery ` , implements #2
The `CalendarQuery ` can be used to fetch calendar data from our `API`.
Example usage:

```swift
let api = SchedJoulesApiClient(accessToken: "YOUR_API_KEY")
        let icsQuery = CalendarQuery(url: "https://iphone.schedjoules.com/calendars/afdd5213056f?l=en&x=6cdd34")
        api.execute(query: icsQuery, completion: { result in
            switch result {
            case let .success(calendar):
                // Do something with the Calendar object, possibly show it in the UI.
            case let .failure(apiError):
                print(apiError)
            }
        })
```
