+++ ApiClient/Tests/ApiClientTests.swift
@@ -102,10 +102,10 @@ class ApiClientTests: XCTestCase {
         waitForExpectations(timeout: 10.0) { (_) -> Void in }
     }
     
-    func testIcsQuery() {
+    func testCalendarQuery() {
         let api = SchedJoulesApiClient(accessToken: getApiKey())
         let responseExpectation = expectation(description: "Received response")
-        let icsQuery = IcsQuery(url: "https://iphone.schedjoules.com/calendars/afdd5213056f?l=en&x=6cdd34")
+        let icsQuery = CalendarQuery(url: "https://iphone.schedjoules.com/calendars/afdd5213056f?l=en&x=6cdd34")
         api.execute(query: icsQuery, completion: { result in
             switch result {
             case let .success(calendar):
